requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
opencv-python>=4.8.0
jinja2>=3.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
